api/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
email-validator==2.1.0.post1
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
sqlalchemy[asyncio]>=2.0.0

# Machine Learning & Privacy
torch==2.1.1
torchvision==0.16.1
numpy==1.26.2
pillow==10.1.0
tenseal==0.3.14
opacus==1.4.1

# Security
cryptography==41.0.5
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
loguru==0.7.2
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0 